+++ src/fwcloud-exporter/database-importer/terraformer/terraformer.ts
@@ -54,18 +54,9 @@ export class Terraformer {
      * 
      * @param exportResults 
      */
-    public async terraform(exportResults: ExporterResult): Promise<ExporterResult> {
-        const result: ExporterResult = new ExporterResult();
-
-        const data: ExporterResultData = exportResults.getAll();
-        
-        for(let tableName in data) {
-            const terraformer: TableTerraformer = await (await this.getTerraformer(tableName)).make(this._mapper, this._queryRunner, this.eventEmitter);
-            const terraformedData: Array<object> = await terraformer.terraform(tableName, data[tableName]);
-            result.addTableData(tableName, terraformedData);
-        }
-
-        return result;
+    public async terraform(tableName: string, data: object[]): Promise<object[]> {
+        const terraformer: TableTerraformer = await (await this.getTerraformer(tableName)).make(this._mapper, this._queryRunner, this.eventEmitter);
+        return await terraformer.terraform(tableName, data);
     }
 
     protected async getTerraformer(tableName: string): Promise<typeof TableTerraformer> {